+++ Runtime/Scripts/Components/TetraGenChunk.cs
@@ -72,7 +72,9 @@ namespace TetraGen
         /// <summary> a single Vector3 must be passed to the GPU as a float array </summary>
         private readonly float[] gpuIn_cellScale = new float[3];
         /// <summary> calculate chunk bounds in GenerationStart</summary>
-        private Bounds bounds;
+        private Bounds chunkBounds;
+        /// <summary> Store discarded mesh prefab instances in a pool rather than destroying them</summary>
+        private readonly Stack<GameObject> meshObjectPool = new Stack<GameObject>();
 
         private void OnDestroy()
         {
@@ -85,19 +87,19 @@ namespace TetraGen
             int cellDataLength = cellCount.x * cellCount.y * cellCount.z;
             int triangleDataLength = cellDataLength * meshKernel.trianglesPerCell;
             int weightBufferLength = (cellCount.x + 1) * (cellCount.y + 1) * (cellCount.z + 1);
-            Vector3 chunkBounds;
+            Vector3 chunkBoundSize;
 
-            chunkBounds.x = cellCount.x * cellScale.x;
-            chunkBounds.y = cellCount.x * cellScale.x;
-            chunkBounds.z = cellCount.x * cellScale.x;
-            bounds = new Bounds(chunkBounds / 2, chunkBounds);
+            chunkBoundSize.x = cellCount.x * cellScale.x;
+            chunkBoundSize.y = cellCount.y * cellScale.y;
+            chunkBoundSize.z = cellCount.z * cellScale.z;
 
+            chunkBounds = new Bounds(chunkBoundSize / 2, chunkBoundSize);
             triangleData = new Triangle[triangleDataLength];
             tCountData = new int[cellDataLength];
             vertexData = new Vertex[verticesPerMesh];
             triangleMap = new ushort[verticesPerMesh];
 
-            for (int t = 0; t < verticesPerMesh; t++)
+            for (int t = 0; t < triangleMap.Length; t++)
                 triangleMap[t] = (ushort)t;
 
             weightBuffer = new ComputeBuffer(weightBufferLength, sizeof(float) * 5);
@@ -112,11 +114,11 @@ namespace TetraGen
             generationReady = true;
         }
 
-        private readonly Queue<TetraGenShape> shapeQueue = new Queue<TetraGenShape>();
+        //private readonly Queue<TetraGenShape> shapeQueue = new Queue<TetraGenShape>();
         /// <summary> Pupulates the triangleData array for mesh generation. </summary>
         /// <param name="shapes"> shape data passed to the GPU to form the signed distance field </param>
         /// <param name="chunk2world"> transforms chunk space into world space on the GPU</param>
-        public void SetTriangleBuffer(TetraGenShape[] shapes, Transform master, Vector3Int chunkIndex)
+        public void SetTriangleBuffer(List<TetraGenShape> shapes, Transform master, Vector3Int chunkIndex)
         {
             if (!generationReady)
                 GenerationStart();
@@ -125,17 +127,11 @@ namespace TetraGen
             chunkCenter.x = chunkIndex.x * cellScale.x;
             chunkCenter.y = chunkIndex.y * cellScale.y;
             chunkCenter.z = chunkIndex.z * cellScale.z;
-
-            for (int i = 0, ie = shapes.Length; i < ie; i++)
-                //if (shapes[i].activeRadius <= 0 || Vector3.Distance(shapes[i].transform.position, chunkCenter) <= shapes[i].activeRadius)
-                    shapeQueue.Enqueue(shapes[i]);
-
             Vector3 masterPosition;
             masterPosition.x = cellScale.x * cellCount.x * chunkIndex.x;
             masterPosition.y = cellScale.y * cellCount.y * chunkIndex.y;
             masterPosition.z = cellScale.z * cellCount.z * chunkIndex.z;
             masterPosition = master.TransformPoint(masterPosition);
-
             Matrix4x4 chunk2world = Matrix4x4.TRS(masterPosition, master.rotation, master.lossyScale);
             Matrix4x4 world2Master = master.worldToLocalMatrix;
 
@@ -155,8 +151,7 @@ namespace TetraGen
                 cellCount.y + 1,
                 cellCount.z + 1);
 
-            //for (int i = 0, ie = shapes.Length; i < ie; i++)
-            foreach( TetraGenShape shape in shapeQueue)
+            foreach( TetraGenShape shape in shapes)
             {
                 // SHAPE BUFFER
                 gpuIn_shapeData[0] = shape.ToShapeData();
@@ -189,8 +184,6 @@ namespace TetraGen
                     cellCount.z + 1);
             }
 
-            shapeQueue.Clear();
-
             // MESH KERNEL
             meshKernel.computer.SetBuffer(meshKernel.id, "weightBuffer", weightBuffer);
             meshKernel.computer.SetBuffer(meshKernel.id, "tBuffer", triangleBuffer);
@@ -207,9 +200,9 @@ namespace TetraGen
 
         /// <summary> Forms meshes out of triangle data generated in SetTriangleBuffer. </summary>
         /// <remarks> Make sure SetTriangleBuffer has ran before executing this method.</remarks>
-        /// <param name="meshSet"> a list of pre-existing GameObjects with meshes for reuse to avoid excessive instantiations </param>
+        /// <param name="chunkMeshes"> a list of pre-existing GameObjects with meshes for reuse to avoid excessive instantiations </param>
         /// <returns> a list of the meshes that were generated </returns>
-        public List<GameObject> GenerateMeshes(Transform parent, Vector3 position, Quaternion rotation, List<GameObject> meshSet = null)
+        public void GenerateMeshes(Transform parent, Vector3 position, Quaternion rotation, List<GameObject> chunkMeshes)
         {
             int tCount = 0;
             int meshCount;
@@ -225,24 +218,39 @@ namespace TetraGen
 
             // Prepare mesh list
             meshCount = Mathf.CeilToInt((float)tCount / trianglesPerMesh);
-            meshSet = meshSet ?? new List<GameObject>(meshCount);
+            chunkMeshes = chunkMeshes ?? new List<GameObject>(meshCount);
 
-            for (int i = meshCount; i < meshSet.Count; i++)
-                DestroyImmediate(meshSet[i]);
+            // Trim excess mesh GameObjects
+            for (int i = meshCount; i < chunkMeshes.Count; i++)
+            {
+                chunkMeshes[i].TryGetComponent(out MeshFilter mf);
+                chunkMeshes[i].SetActive(false);
+                mf.sharedMesh.Clear();
+                meshObjectPool.Push(chunkMeshes[i]);
+            }
 
-            if (meshSet.Count > meshCount)
-                meshSet.RemoveRange(meshCount, meshSet.Count - meshCount);
-
-            for (int i = meshCount - meshSet.Count; i > 0; i--)
-                meshSet.Add(Instantiate(meshTemplate.gameObject, parent));
+            // Add or remove unused indeces
+            if (chunkMeshes.Count > meshCount)
+                chunkMeshes.RemoveRange(meshCount, chunkMeshes.Count - meshCount);
+            else
+                for (int i = meshCount - chunkMeshes.Count; i > 0; i--)
+                    if (meshObjectPool.Count != 0)
+                    {
+                        GameObject m = meshObjectPool.Pop();
+                        m.SetActive(true);
+                        chunkMeshes.Add(m);
+                    }
+                    else
+                        chunkMeshes.Add(Instantiate(meshTemplate.gameObject, parent));
 
             // Mesh Generation
             for (int m = 0; m < meshCount; m++)
             {
-                GameObject mo = meshSet[m];
-                MeshRenderer mr = mo.GetComponent<MeshRenderer>();
-                MeshFilter mf = mo.GetComponent<MeshFilter>();
-                MeshCollider mc = mo.GetComponent<MeshCollider>();
+                GameObject mo = chunkMeshes[m];
+                mo.TryGetComponent(out Transform tr);
+                mo.TryGetComponent(out MeshRenderer mr);
+                mo.TryGetComponent(out MeshFilter mf);
+                mo.TryGetComponent(out MeshCollider mc);
                 Mesh mesh = mf.sharedMesh ? mf.sharedMesh : new Mesh();
                 int tStart = m * trianglesPerMesh;
                 int tEnd = Mathf.Min(tCount, tStart + trianglesPerMesh);
@@ -262,10 +270,8 @@ namespace TetraGen
                 mesh.SetVertexBufferParams(vCount, Vertex.meshAttributes);
                 mesh.SetVertexBufferData(vertexData, 0, 0, vCount);
                 mesh.SetTriangles(triangleMap, 0, vCount, 0, false, 0);
-                mesh.bounds = bounds;
-                mesh.name = mo.name = "Generated Mesh " + m;
-                mo.transform.position = position;
-                mo.transform.rotation = rotation;
+                mesh.bounds = chunkBounds;
+                tr.SetPositionAndRotation(position, rotation);
                 mo.layer = parent.gameObject.layer;
                 mf.sharedMesh = mesh;
 
@@ -275,8 +281,6 @@ namespace TetraGen
                 if (generateCollision && mc && mc.enabled)
                     mc.sharedMesh = mesh;
             }
-
-            return meshSet;
         }
 
         /// <summary> Frees all memory associated with mesh generation </summary>
@@ -295,39 +299,5 @@ namespace TetraGen
                 triangleMap = null;
             }
         }
-
-        /*//
-        private void OnDrawGizmosSelected()
-        {
-            Transform cam = Camera.current.transform;
-            TetraGenMaster master = GetComponentInParent<TetraGenMaster>();
-
-            for (int x = 0; x < master.cellCount.x; x++)
-                for (int y = 0; y < master.cellCount.y; y++)
-                    for (int z = 0; z < master.cellCount.z; z++)
-                    {
-                        Vector3 localPos = new Vector3(
-                            x * master.cellScale.x,
-                            y * master.cellScale.y,
-                            z * master.cellScale.z);
-                        Vector3 worldPos = transform.TransformPoint(localPos);
-                        float camDistance = Vector3.Distance(worldPos, cam.position);
-                        float distFactor = InvLerp(0, 4, camDistance);
-                        float angle = Vector3.Angle(cam.forward, worldPos - cam.position);
-
-                        if (angle < 45 && camDistance < 4 && camDistance > 0)
-                        {
-                            Gizmos.color = new Color(1, 0, 0, distFactor);
-                            Gizmos.DrawLine(worldPos, transform.TransformPoint(localPos + new Vector3(master.cellScale.x, 0, 0)));
-                            Gizmos.color = new Color(0, 1, 0, distFactor);
-                            Gizmos.DrawLine(worldPos, transform.TransformPoint(localPos + new Vector3(0, master.cellScale.y, 0)));
-                            Gizmos.color = new Color(0, 0, 1, distFactor);
-                            Gizmos.DrawLine(worldPos, transform.TransformPoint(localPos + new Vector3(0, 0, master.cellScale.z)));
-                            Gizmos.color = new Color(x % 2, y % 2, z % 2, distFactor);
-                            Gizmos.DrawCube(worldPos + transform.TransformVector(master.cellScale) * 0.05f, master.cellScale * 0.1f);
-                        }
-                    }
-        }
-        //*/
     }
 }