+++ Runtime/Scripts/Components/TetraGenMaster.cs
@@ -8,7 +8,6 @@ namespace TetraGen
     ///     Main class of the TetraGen system.
     ///     Place anywhere in your scene to act as the origin of an isosurface.
     /// </summary>
-    [ExecuteInEditMode]
     [DisallowMultipleComponent]
     public class TetraGenMaster : MonoBehaviour
     {
@@ -19,38 +18,40 @@ namespace TetraGen
         [Tooltip("Number of chunks to generate per direction")]
         public Vector3Int chunkCount = new Vector3Int(8, 8, 8);
         [Tooltip("The transform that contains the TetraGenShapes that will form the isosurface mesh.")]
-        public Transform shapeContainer;
+        public TGShapeContainer shapeContainer;
 
         /// <summary> TetraGenChunk handles the mesh generation </summary>
         private TetraGenChunk chunk;
         /// <summary> true when this TetraGenMaster has been initialized for generation </summary>
         private bool readyToGenerate = false;
         /// <summary> Storage of which meshes are associated with which chunk coordinates </summary>
-        private readonly Dictionary<Vector3Int, List<GameObject>> meshSets = new Dictionary<Vector3Int, List<GameObject>>();
+        private readonly Dictionary<Vector3Int, ChunkData> activeChunks = new Dictionary<Vector3Int, ChunkData>();
+        /// <summary> This is the target followed in terrain mode </summary>
+        private Transform followTarget = null;
+        private Vector3Int previousTargetChunk = Vector3Int.zero;
+        private Vector3Int[] chunkLoadOrder = null;
+        private int chunkLoadCount = 0;
 
         private void OnValidate()
         {
             if(generationType.Equals(GenerationType.Static))
-                Close();
+                GenerationEnd();
         }
 
         private void Awake()
         {
-            chunk = GetComponent<TetraGenChunk>();
+            TryGetComponent(out chunk);
         }
 
         private void OnEnable()
         {
             if (!generationType.Equals(GenerationType.Static))
                 GenStart();
-
-            if (!shapeContainer)
-                shapeContainer = transform;
         }
 
         private void OnDisable()
         {
-            Close();
+            GenerationEnd();
         }
 
         private void Update()
@@ -60,9 +61,9 @@ namespace TetraGen
                 case GenerationType.Static:
                     break;
                 case GenerationType.Realtime:
-                    RealtimeUpdate(); break;
+                    GenerateRealtime(); break;
                 case GenerationType.Terrain:
-                    FollowTarget(); break;
+                    GenerateFollowTarget(); break;
             }
         }
 
@@ -72,7 +73,7 @@ namespace TetraGen
             foreach (Component mesh in GetComponentsInChildren<MeshFilter>())
                 DestroyImmediate(mesh.gameObject);
 
-            meshSets.Clear();
+            activeChunks.Clear();
         }
 
         /// <summary>
@@ -81,8 +82,8 @@ namespace TetraGen
         public void Generate()
         {
             GenStart();
-            RealtimeUpdate();
-            Close();
+            GenerateRealtime();
+            GenerationEnd();
         }
 
         /// <summary>
@@ -95,34 +96,74 @@ namespace TetraGen
             chunk.GenerationStart();
 
             readyToGenerate = true;
+
+            if (generationType.Equals(GenerationType.Terrain))
+            {
+                SetTargetToCamera();
+
+                SortedList<float, Vector3Int> sortedChunks = new SortedList<float, Vector3Int>();
+                Vector3Int centerChunk = chunkCount / 2;
+                for (int x = 0; x < chunkCount.x; x++)
+                    for (int y = 0; y < chunkCount.y; y++)
+                        for (int z = 0; z < chunkCount.z; z++)
+                        {
+                            Vector3Int offset = new Vector3Int(
+                                x - centerChunk.x,
+                                y - centerChunk.y,
+                                z - centerChunk.z);
+                            float dist = offset.sqrMagnitude;
+
+                            while (sortedChunks.ContainsKey(dist))
+                                dist += 0.0001f;
+
+                            sortedChunks.Add(dist, offset);
+                        }
+                chunkLoadOrder = sortedChunks.Values.ToArray();
+                chunkLoadCount = 0;
+            }
+        }
+
+        /// <summary>
+        /// Sets the follow target to the main camera.
+        /// </summary>
+        public void SetTargetToCamera()
+        {
+#if UNITY_EDITOR
+            if (UnityEditor.EditorApplication.isPlaying)
+                followTarget = Camera.main.transform;
+            else
+                followTarget = UnityEditor.SceneView.GetAllSceneCameras()[0].transform;
+#else
+            followTarget = Camera.main.transform;
+#endif
         }
 
         /// <summary>
         /// Realtime mesh generation method. Updates every chunk every frame.
         /// </summary>
-        private void RealtimeUpdate()
+        private void GenerateRealtime()
         {
             if (!readyToGenerate)
                 GenStart();
-
-            TetraGenShape[] shapes = TetraGenShape.GetShapesIn(shapeContainer);
 
             for (int x = 0; x < chunkCount.x; x++)
                 for (int y = 0; y < chunkCount.y; y++)
                     for (int z = 0; z < chunkCount.z; z++)
                     {
-                        Vector3Int key = new Vector3Int(x, y, z);
-
-                        if (meshSets.TryGetValue(key, out List<GameObject> currentMeshes))
-                            meshSets.Remove(key);
-
+                        Vector3Int chunkId = new Vector3Int(x, y, z);
                         Vector3 position = transform.TransformPoint(new Vector3(
-                            key.x * chunk.cellScale.x * chunk.cellCount.x,
-                            key.y * chunk.cellScale.y * chunk.cellCount.y,
-                            key.z * chunk.cellScale.z * chunk.cellCount.z));
+                            chunkId.x * chunk.cellScale.x * chunk.cellCount.x,
+                            chunkId.y * chunk.cellScale.y * chunk.cellCount.y,
+                            chunkId.z * chunk.cellScale.z * chunk.cellCount.z));
 
-                        chunk.SetTriangleBuffer(shapes, transform, key);
-                        meshSets.Add(key, chunk.GenerateMeshes(transform, position, transform.rotation, currentMeshes));
+                        if (!activeChunks.TryGetValue(chunkId, out ChunkData currentChunk))
+                        {
+                            currentChunk = new ChunkData();
+                            activeChunks.Add(chunkId, currentChunk);
+                        }
+
+                        chunk.SetTriangleBuffer(shapeContainer.shapes, transform, chunkId);
+                        chunk.GenerateMeshes(transform, position, transform.rotation, currentChunk.meshes);
                     }
         }
 
@@ -130,108 +171,96 @@ namespace TetraGen
         /// Terrain mesh generation method. Updates chunks closest to the camera, one chunk per frame.
         /// Swaps out far chunks for close chunks to maintain a stable amount of geometry.
         /// </summary>
-        private void FollowTarget()
+        private void GenerateFollowTarget()
         {
             if (!readyToGenerate)
                 GenStart();
 
-            Vector3 target;
+            Vector3 targetPosition = followTarget ? transform.InverseTransformPoint(followTarget.position) : Vector3.zero;
+            Vector3Int targetChunk = new Vector3Int(
+                Mathf.FloorToInt(targetPosition.x / chunk.cellScale.x / chunk.cellCount.x),
+                Mathf.FloorToInt(targetPosition.y / chunk.cellScale.y / chunk.cellCount.y),
+                Mathf.FloorToInt(targetPosition.z / chunk.cellScale.z / chunk.cellCount.z));
 
-#if UNITY_EDITOR
-            if(UnityEditor.EditorApplication.isPlaying)
-                target = transform.InverseTransformPoint(Camera.main.transform.position);
-            else
-                target = transform.InverseTransformPoint(UnityEditor.SceneView.GetAllSceneCameras()[0].transform.position);
-#else
-            target = transform.InverseTransformPoint(Camera.main.transform.position);
-#endif
+            // Closest unloaded in-bounds chunk
+            Vector3Int closestEmptyChunkId;
+            float closestDistance;
 
-            Vector3Int currentCell = new Vector3Int(
-                Mathf.FloorToInt(target.x / chunk.cellScale.x / chunk.cellCount.x),
-                Mathf.FloorToInt(target.y / chunk.cellScale.y / chunk.cellCount.y),
-                Mathf.FloorToInt(target.z / chunk.cellScale.z / chunk.cellCount.z));
-            Vector3Int minCell = new Vector3Int(
-                currentCell.x - chunkCount.x / 2,
-                currentCell.y - chunkCount.y / 2,
-                currentCell.z - chunkCount.z / 2);
-            Vector3Int maxCell = new Vector3Int(
-                currentCell.x + chunkCount.x / 2 + chunkCount.x % 2,
-                currentCell.y + chunkCount.y / 2 + chunkCount.y % 2,
-                currentCell.z + chunkCount.z / 2 + chunkCount.z % 2);
+            if (!targetChunk.Equals(previousTargetChunk))
+            {
+                chunkLoadCount = 0;
+                previousTargetChunk = targetChunk;
+            }
 
-            // Get closest empty space
-            Vector3Int closestEmpty = default;
-            float closestDistance = float.MaxValue;
-            for (int x = minCell.x; x < maxCell.x; x++)
-                for (int y = minCell.y; y < maxCell.y; y++)
-                    for (int z = minCell.z; z < maxCell.z; z++)
-                    {
-                        Vector3Int checkCell = new Vector3Int(x, y, z);
-                        float distance = (
-                            new Vector3(x, y, z) -
-                            new Vector3(currentCell.x, currentCell.y, currentCell.z))
-                            .sqrMagnitude;
+            while (chunkLoadCount < chunkLoadOrder.Length && activeChunks.ContainsKey(targetChunk + chunkLoadOrder[chunkLoadCount]))
+                chunkLoadCount++;
 
-                        if (distance < closestDistance && !meshSets.ContainsKey(checkCell))
-                        {
-                            closestEmpty = checkCell;
-                            closestDistance = distance;
-                        }
-                    }
+            if (chunkLoadCount < chunkLoadOrder.Length)
+            {
+                closestEmptyChunkId = targetChunk + chunkLoadOrder[chunkLoadCount];
+                closestDistance = 0;
+            }
+            else return;
 
-            /* Get farthest loaded mesh set.
-             * Only select from out of range chunks */
-            List<GameObject> farthestSet = null;
-            Vector3Int farthestKey = default;
+            // Farthest loaded out-of-bounds chunk
+            ChunkData farthestActiveChunk = null;
+            Vector3Int farthestActiveChunkId = default;
             float farthestDistance = -1f;
-            foreach (var kvp in meshSets)
+            Vector3Int minCell = new Vector3Int(
+                targetChunk.x - chunkCount.x / 2,
+                targetChunk.y - chunkCount.y / 2,
+                targetChunk.z - chunkCount.z / 2);
+            Vector3Int maxCell = new Vector3Int(
+                targetChunk.x + chunkCount.x / 2 + chunkCount.x % 2,
+                targetChunk.y + chunkCount.y / 2 + chunkCount.y % 2,
+                targetChunk.z + chunkCount.z / 2 + chunkCount.z % 2);
+
+            foreach (var kvp in activeChunks)
                 if (kvp.Key.x < minCell.x || kvp.Key.x >= maxCell.x
                     || kvp.Key.y < minCell.y || kvp.Key.y >= maxCell.y
                     || kvp.Key.z < minCell.z || kvp.Key.z >= maxCell.z)
                 {
-                    float distance = (
-                        new Vector3(kvp.Key.x, kvp.Key.y, kvp.Key.z) -
-                        new Vector3(currentCell.x, currentCell.y, currentCell.z))
-                        .sqrMagnitude;
+                    float distance = (kvp.Key - targetChunk).sqrMagnitude;
 
                     if (distance > farthestDistance)
                     {
-                        farthestSet = kvp.Value;
-                        farthestKey = kvp.Key;
+                        farthestActiveChunk = kvp.Value;
+                        farthestActiveChunkId = kvp.Key;
                         farthestDistance = distance;
                     }
                 }
 
-            /* Move farthest mesh set to the closest empty space.
-               Reuse them to avoid unnecessary instantiations. */
+            // Unload far chunk, Load closest chunk
             if (closestDistance != float.MaxValue)
             {
-                TetraGenShape[] shapes = TetraGenShape.GetShapesIn(shapeContainer);
+                Vector3 chunkPosition = transform.TransformPoint(new Vector3(
+                    closestEmptyChunkId.x * chunk.cellScale.x * chunk.cellCount.x,
+                    closestEmptyChunkId.y * chunk.cellScale.y * chunk.cellCount.y,
+                    closestEmptyChunkId.z * chunk.cellScale.z * chunk.cellCount.z));
 
-                if (farthestSet != null)
-                    meshSets.Remove(farthestKey);
+                if (farthestActiveChunk != null)
+                    activeChunks.Remove(farthestActiveChunkId);
+                else
+                    farthestActiveChunk = new ChunkData();
 
-                Vector3 position = transform.TransformPoint(new Vector3(
-                    closestEmpty.x * chunk.cellScale.x * chunk.cellCount.x,
-                    closestEmpty.y * chunk.cellScale.y * chunk.cellCount.y,
-                    closestEmpty.z * chunk.cellScale.z * chunk.cellCount.z));
-
-                chunk.SetTriangleBuffer(shapes, transform, closestEmpty);
-                meshSets.Add(closestEmpty, chunk.GenerateMeshes(transform, position, transform.rotation, farthestSet));
+                chunk.SetTriangleBuffer(shapeContainer.shapes, transform, closestEmptyChunkId);
+                chunk.GenerateMeshes(transform, chunkPosition, transform.rotation, farthestActiveChunk.meshes);
+                activeChunks.Add(closestEmptyChunkId, farthestActiveChunk);
             }
         } 
 
         /// <summary>
         /// Frees up all memory involved in mesh generation.
         /// </summary>
-        private void Close()
+        private void GenerationEnd()
         {
             if (!readyToGenerate)
                 return;
 
             chunk.GenerationEnd();
-            meshSets.Clear();
+            activeChunks.Clear();
             readyToGenerate = false;
+            chunkLoadOrder = null;
         }
     }
 }