+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## [Unreleased]
+## Unreleased
 
-## [0.1.0] - 2020-10-09
+## 0.2.0 - 2020-11-24
+- Shape containers now use the designated class TGShapeContainer
+- Less memory allocation all around.
+- Added the ChunkData class to store more info about chunks later on.
+- More efficient management of chunk.
+
+## 0.1.0 - 2020-10-09
 ### Added
 - Finally turned this project into a package!